+++ src/main/java/com/example/wpct/controller/sysUser/RoleApi.java
@@ -1,10 +1,14 @@
 package com.example.wpct.controller.sysUser;
 
 import com.example.wpct.entity.Role;
+import com.example.wpct.entity.SysUser;
 import com.example.wpct.service.RoleService;
 import com.example.wpct.utils.ResultBody;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.SneakyThrows;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -49,9 +53,19 @@ public class RoleApi {
         roleService.update(id,name,permission);
         return ResultBody.ok(null);
     }
+    @SneakyThrows
     @ApiOperation("删除角色")
     @RequestMapping(value = "/delete",method = RequestMethod.GET)
     public Object delete(@RequestParam int id){
+        Subject subject = SecurityUtils.getSubject();
+        SysUser user = (SysUser)subject.getPrincipal();
+        if (user == null){
+            throw new Exception("请先登录");
+        }
+        String permission = roleService.getById(user.getRole()).getPermission();
+        if (!permission.contains("删除")) {
+            return ResultBody.ok("您没有删除权限");
+        }
         roleService.delete(id);
         return ResultBody.ok(null);
     }
